requirements.txt
openai
python-dotenv
canvasapi
